Operation latency is now stored in a decimal (we need fractions, and precise ones)
--HG--
branch : feature/fsm-transformer

+++ Hast.Synthesis/IDeviceDriver.cs
@@ -16,6 +16,6 @@ namespace Hast.Synthesis
     {
         IDeviceManifest DeviceManifest { get; }
 
-        uint GetClockCyclesNeededForOperation(BinaryOperatorType operation);
+        decimal GetClockCyclesNeededForOperation(BinaryOperatorType operation);
     }
 }
